+++ wolfssl/internal.h
@@ -2294,6 +2294,17 @@ WOLFSSL_LOCAL int  HandleTlsResumption(WOLFSSL* ssl, Suites* clSuites);
 #ifdef WOLFSSL_TLS13
 WOLFSSL_LOCAL byte SuiteMac(const byte* suite);
 #endif
+#if (defined(WOLFSSL_DTLS) || defined(WOLFSSL_SEND_HRR_COOKIE)) && \
+    !defined(NO_WOLFSSL_SERVER)
+WOLFSSL_LOCAL void FreeCookieSecret(WOLFSSL* ssl, buffer* secret);
+WOLFSSL_LOCAL int SetCookieSecret(WOLFSSL* ssl, buffer* dst,
+                                  const byte* secret, word32 secretSz,
+                                  const char* name);
+WOLFSSL_LOCAL int CookiePolicySet(WOLFSSL* ssl, const byte* hrrSecret,
+                                  word32 hrrSecretSz, int replaceHrr);
+WOLFSSL_LOCAL int CookiePolicyEnable(WOLFSSL* ssl);
+WOLFSSL_LOCAL int CheckCookieState(WOLFSSL* ssl);
+#endif
 WOLFSSL_LOCAL int  DoClientHello(WOLFSSL* ssl, const byte* input, word32* inOutIdx,
                              word32 helloSz);
 #ifdef WOLFSSL_TLS13
@@ -5617,8 +5628,9 @@ struct Options {
     word16            verifyPostHandshake:1; /* Only send client cert req post
                                               * handshake, not also during */
 #endif
-#if defined(WOLFSSL_TLS13) && !defined(NO_WOLFSSL_SERVER)
-    word16            sendCookie:1;       /* Server creates a Cookie in HRR */
+#if (defined(WOLFSSL_TLS13) || defined(WOLFSSL_DTLS)) && \
+    !defined(NO_WOLFSSL_SERVER)
+    word16            sendCookie:1;       /* DTLS cookies / TLS 1.3 HRR cookie */
 #endif
 #ifdef WOLFSSL_ALT_CERT_CHAINS
     word16            usingAltCertChain:1;/* Alternate cert chain was used */
@@ -7961,6 +7973,9 @@ WOLFSSL_LOCAL int DtlsUpdateWindow(WOLFSSL* ssl);
 WOLFSSL_LOCAL void DtlsResetState(WOLFSSL *ssl);
 WOLFSSL_LOCAL int DtlsIgnoreError(int err);
 WOLFSSL_LOCAL void DtlsSetSeqNumForReply(WOLFSSL* ssl);
+#ifndef NO_WOLFSSL_SERVER
+WOLFSSL_LOCAL int DtlsNoCookieChGood(WOLFSSL* ssl);
+#endif
 #endif
 
 #ifdef WOLFSSL_DTLS13